backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3
psycopg2-binary>=2.9.0
python-decouple==3.8
Pillow>=10.1.0
python-dotenv==1.0.0

# Satellite Imagery & Tree Detection
rasterio>=1.3.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
shapely>=2.0.0

# LiDAR Processing
laspy[lazrs]>=2.4.0
pyproj>=3.5.0

# Async Task Processing
celery>=5.3.0
redis>=4.5.0